requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pikepdf
python-dotenv
